requirements.txt
astroid==1.5.3
click==6.7
Django==1.11.7
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
PyJWT==1.5.3
pylint==1.7.4
python-dotenv==0.7.1
pytz==2017.3
six==1.11.0
wrapt==1.10.11
